Trabajando en el modulo empleados

+++ app/src/main/java/com/Softpig/Model/Employee.java
@@ -50,7 +50,7 @@ public class Employee implements Serializable {
         this.salary = salary;
     }
 
-    public int getIdEmployee() {
+    public short getIdEmployee() {
         return idEmployee;
     }
